get clang build working again

+++ src/scrabble/primes.cpp
@@ -11,7 +11,7 @@ std::array<uint64_t, 32> Primes::FirstNPrimes(int n) {
   int i = 1;
   for (int x = 3;; x += 2) {
     bool is_prime = true;
-    for (int d_i = 1; ret[d_i] * ret[d_i] <= x; d_i++) {
+    for (int d_i = 1; static_cast<int>(ret[d_i] * ret[d_i]) <= x; d_i++) {
       if (x % ret[d_i] == 0) {
         is_prime = false;
         break;
